cluster-watcher/cloud_function_source/requirements.txt
functions-framework==3.*
google-cloud-build
google-cloud-edgecontainer
google-cloud-edgenetwork
google-cloud-storage
